+++ contracts/nft/erc721m/ERC721CM.sol
@@ -76,7 +76,7 @@ contract ERC721CM is IERC721M, ERC721ACQueryable, Ownable, ReentrancyGuard, Cosi
         uint256 maxMintableSupply,
         uint256 globalWalletLimit,
         address cosigner,
-        uint64 timestampExpirySeconds,
+        uint256 timestampExpirySeconds,
         address mintCurrency,
         address fundReceiver
     ) Ownable(msg.sender) ERC721ACQueryable(collectionName, collectionSymbol) {
@@ -308,7 +308,7 @@ contract ERC721CM is IERC721M, ERC721ACQueryable, Ownable, ReentrancyGuard, Cosi
      * timestamp - the current timestamp
      * signature - the signature from cosigner if using cosigner.
      */
-    function mint(uint32 qty, uint32 limit, bytes32[] calldata proof, uint64 timestamp, bytes calldata signature)
+    function mint(uint32 qty, uint32 limit, bytes32[] calldata proof, uint256 timestamp, bytes calldata signature)
         external
         payable
         virtual
@@ -326,7 +326,7 @@ contract ERC721CM is IERC721M, ERC721ACQueryable, Ownable, ReentrancyGuard, Cosi
      * timestamp - the current timestamp
      * signature - the signature from cosigner if using cosigner.
      */
-    function crossmint(uint32 qty, address to, bytes32[] calldata proof, uint64 timestamp, bytes calldata signature)
+    function crossmint(uint32 qty, address to, bytes32[] calldata proof, uint256 timestamp, bytes calldata signature)
         external
         payable
         nonReentrant
@@ -354,7 +354,7 @@ contract ERC721CM is IERC721M, ERC721ACQueryable, Ownable, ReentrancyGuard, Cosi
         address to,
         uint32 limit,
         bytes32[] calldata proof,
-        uint64 timestamp,
+        uint256 timestamp,
         bytes calldata signature
     ) external payable onlyAuthorizedMinter {
         _mintInternal(qty, to, limit, proof, timestamp, signature);
@@ -368,10 +368,10 @@ contract ERC721CM is IERC721M, ERC721ACQueryable, Ownable, ReentrancyGuard, Cosi
         address to,
         uint32 limit,
         bytes32[] calldata proof,
-        uint64 timestamp,
+        uint256 timestamp,
         bytes calldata signature
     ) internal canMint hasSupply(qty) {
-        uint64 stageTimestamp = uint64(block.timestamp);
+        uint256 stageTimestamp = block.timestamp;
         bool waiveMintFee = false;
 
         if (getCosigner() != address(0)) {
@@ -514,7 +514,7 @@ contract ERC721CM is IERC721M, ERC721ACQueryable, Ownable, ReentrancyGuard, Cosi
     /**
      * @dev Returns the current active stage based on timestamp.
      */
-    function getActiveStageFromTimestamp(uint64 timestamp) public view returns (uint256) {
+    function getActiveStageFromTimestamp(uint256 timestamp) public view returns (uint256) {
         for (uint256 i = 0; i < _mintStages.length;) {
             if (timestamp >= _mintStages[i].startTimeUnixSeconds && timestamp < _mintStages[i].endTimeUnixSeconds) {
                 return i;
@@ -529,7 +529,7 @@ contract ERC721CM is IERC721M, ERC721ACQueryable, Ownable, ReentrancyGuard, Cosi
     /**
      * @dev Validates the start timestamp is before end timestamp. Used when updating stages.
      */
-    function _assertValidStartAndEndTimestamp(uint64 start, uint64 end) internal pure {
+    function _assertValidStartAndEndTimestamp(uint256 start, uint256 end) internal pure {
         if (start >= end) revert InvalidStartAndEndTimestamp();
     }
