+++ contracts/nft/erc721m/ERC721MErrorsAndEvents.sol
@@ -11,8 +11,8 @@ interface ERC721MErrorsAndEvents is ErrorsAndEvents {
         uint32 walletLimit,
         bytes32 merkleRoot,
         uint24 maxStageSupply,
-        uint64 startTimeUnixSeconds,
-        uint64 endTimeUnixSeconds
+        uint256 startTimeUnixSeconds,
+        uint256 endTimeUnixSeconds
     );
 
     event SetMaxMintableSupply(uint256 maxMintableSupply);
