+++ contracts/nft/erc721m/extensions/ERC721MOperatorFilterer.sol
@@ -14,7 +14,7 @@ contract ERC721MOperatorFilterer is ERC721M, UpdatableOperatorFilterer {
         uint256 maxMintableSupply,
         uint256 globalWalletLimit,
         address cosigner,
-        uint64 timestampExpirySeconds,
+        uint256 timestampExpirySeconds,
         address mintCurrency,
         address fundReceiver
     )
