+++ contracts/nft/erc721m/interfaces/IERC721M.sol
@@ -16,11 +16,11 @@ interface IERC721M is IERC721AQueryable, ERC721MErrorsAndEvents {
 
     function getStageInfo(uint256 index) external view returns (MintStageInfo memory, uint32, uint256);
 
-    function mint(uint32 qty, uint32 limit, bytes32[] calldata proof, uint64 timestamp, bytes calldata signature)
+    function mint(uint32 qty, uint32 limit, bytes32[] calldata proof, uint256 timestamp, bytes calldata signature)
         external
         payable;
 
-    function crossmint(uint32 qty, address to, bytes32[] calldata proof, uint64 timestamp, bytes calldata signature)
+    function crossmint(uint32 qty, address to, bytes32[] calldata proof, uint256 timestamp, bytes calldata signature)
         external
         payable;
 
@@ -29,7 +29,7 @@ interface IERC721M is IERC721AQueryable, ERC721MErrorsAndEvents {
         address to,
         uint32 limit,
         bytes32[] calldata proof,
-        uint64 timestamp,
+        uint256 timestamp,
         bytes calldata signature
     ) external payable;
 }
